refactor: Busines-logic
- Reservation + Restaurant

+++ programCode/webApi/Services/ReservationService.cs
@@ -13,24 +13,40 @@ namespace RestaurantReservierung.Services
         List<User> users = new List<User>();
 
         private readonly AppDbContext _context;
+        private readonly UserService _userService;
+        private readonly TableService _tableService;
 
         // The min Timespan how long a reservation needs to be.
         public TimeSpan MinReservationTime { get; } = new TimeSpan(1, 0, 0);
 
-        public ReservationService(AppDbContext context)
+        public ReservationService(AppDbContext context, UserService userService, TableService tableService)
         {
             _context = context;
+            _userService = userService;
+            _tableService = tableService;
         }
 
-        public async Task<bool> ReserveAsync(ReservationFormModel model, Table table, User user)
-        {
-            var existingUser = await _context.Users.FindAsync(user.UserId);
-            if (existingUser == null)
-                throw new Exception("User nicht gefunden");
+        public async Task<bool> ReserveAsync(ReservationFormModel model, int tableId)
+        {        
+            var user = await _userService.GetLoggedInUserAsync() ?? throw new Exception("User nicht gefunden");
+            var table = await _tableService.GetTableByIdAsync(tableId) ?? throw new BadHttpRequestException("The table does not exist." );
+
+            if (model.EndTime <= model.StartTime)
+                throw new BadHttpRequestException("Illegal time interval!" );
+
+            if (!IsGreaterThenMinTimeInterval(model))
+                throw new BadHttpRequestException("The reservation time interval has to be >= " + MinReservationTime.ToString() + "." );
+
+            if (IsInPast(model))
+                throw new BadHttpRequestException("The given Time interval is in the past!" );
+
+            if ((await GetReservationsForTimeIntervalAsync(model, table)).Count > 0)
+                throw new BadHttpRequestException("There already exists a reservation in the given time interval!" );
+
 
             var reservation = new Reservation
             {
-                User = existingUser,
+                User = user,
                 Table = table,
                 StartTime = model.StartTime,
                 EndTime = model.EndTime,
@@ -130,8 +146,27 @@ namespace RestaurantReservierung.Services
             return false;
         }
 
-        public async Task<bool> UpdateReservationAsync(ReservationFormModel model, Reservation reservation)
+        public async Task<bool> UpdateReservationAsync(ReservationFormModel model, int reservationId)
         {
+            var user = await _userService.GetLoggedInUserAsync();
+            var reservation = await GetReservationByIdAsync(reservationId);
+
+            if (user.UserId != reservation.UserId && user.Role != "ADMIN")
+                throw new BadHttpRequestException("You don't have Permissions to change the reservation");
+
+            if (model.EndTime <= model.StartTime)
+                throw new BadHttpRequestException("Illegal time interval!");
+
+            if (!IsGreaterThenMinTimeInterval(model))
+                throw new BadHttpRequestException("The reservation time interval has to be >= " + MinReservationTime.ToString() + "." );
+
+            if (IsInPast(model))
+                throw new BadHttpRequestException("The given Time interval is in the past!");
+
+            if (!await CanUpdateReservationAsync(reservation, model))
+                throw new BadHttpRequestException("There already exists a reservation in the given time interval!");
+
+
             reservation.StartTime = model.StartTime;
             reservation.EndTime = model.EndTime;
             reservation.UpdatedAt = DateTime.Now;
@@ -140,8 +175,16 @@ namespace RestaurantReservierung.Services
 
         }
 
-        public async Task<bool> DeleteReservationAsync(Reservation reservation)
+        public async Task<bool> DeleteReservationAsync(int reservationId)
         {
+            var user = await _userService.GetLoggedInUserAsync();
+
+            var reservation = await GetReservationByIdAsync(reservationId);
+
+            if (user != reservation.User && user.Role != "ADMIN" && user != reservation.Table.Restaurant.User)
+                throw new BadHttpRequestException("Your dont have permissions to perform this action!");
+
+
             _context.Reservations.Remove(reservation);
 
             if (await _context.SaveChangesAsync() > 0)
